requirements.txt
matplotlib==3.1.0
numpy==1.16.3
pillow==6.2.2
pyopencl==2018.2.5
PyQt5==5.12.2
scipy==1.3.0
tqdm==4.32.1
